+++ libs/ymir-core/src/ymir/hw/cdblock/cdblock.cpp
@@ -616,12 +616,13 @@ bool CDBlock::SetupGenericPlayback(uint32 startParam, uint32 endParam, uint16 re
         // Find track containing the requested start frame address
         const uint8 trackIndex = session.FindTrackIndex(frameAddress);
         if (trackIndex != 0xFF) {
+            const uint8 index = session.tracks[trackIndex].FindIndex(frameAddress);
             m_status.statusCode = kStatusCodeSeek;
             m_status.flags = 0x8;     // CD-ROM decoding flag
             m_status.repeatCount = 0; // first repeat
             m_status.controlADR = session.tracks[trackIndex].controlADR;
             m_status.track = trackIndex + 1;
-            m_status.index = 1; // TODO: handle indexes
+            m_status.index = index;
 
             // TODO: delay seek for a realistic amount of time
             if (m_status.controlADR == 0x41) {
@@ -660,7 +661,7 @@ bool CDBlock::SetupGenericPlayback(uint32 startParam, uint32 endParam, uint16 re
         }
         if (endParam == 0) {
             endTrack = session.lastTrackIndex + 1;
-            endIndex = 1;
+            endIndex = session.tracks[session.lastTrackIndex].indices.size() + 1;
         }
 
         devlog::debug<grp::play_init>("Track:Index range {:02d}:{:02d}-{:02d}:{:02d} ", startTrack, startIndex,
@@ -672,13 +673,14 @@ bool CDBlock::SetupGenericPlayback(uint32 startParam, uint32 endParam, uint16 re
         uint8 lastTrack = session.lastTrackIndex + 1;
         startTrack = std::clamp(startTrack, firstTrack, lastTrack);
         endTrack = std::clamp(endTrack, firstTrack, lastTrack);
-        devlog::debug<grp::play_init>("Track range after clamping {:02d}-{:02d}", startTrack, endTrack);
-
-        // TODO: tracks need to store index information
+        startIndex = std::clamp<uint8>(startIndex, 1, session.tracks[startTrack - 1].indices.size() + 1);
+        endIndex = std::clamp<uint8>(endIndex, 1, session.tracks[endTrack - 1].indices.size() + 1);
+        devlog::debug<grp::play_init>("Track:Index range after clamping {:02d}:{:02d}-{:02d}:{:02d}", startTrack,
+                                      startIndex, endTrack, endIndex);
 
         // Play frame address range for the specified tracks
-        m_playStartPos = session.tracks[startTrack - 1].startFrameAddress;
-        m_playEndPos = session.tracks[endTrack - 1].endFrameAddress;
+        m_playStartPos = session.tracks[startTrack - 1].indices[startIndex - 1].startFrameAddress;
+        m_playEndPos = session.tracks[endTrack - 1].indices[endIndex - 1].endFrameAddress;
 
         uint32 frameAddress = m_status.frameAddress;
         if (resetPos || frameAddress < m_playStartPos || frameAddress > m_playEndPos) {
@@ -693,7 +695,7 @@ bool CDBlock::SetupGenericPlayback(uint32 startParam, uint32 endParam, uint16 re
             m_status.repeatCount = 0; // first repeat
             m_status.controlADR = track->controlADR;
             m_status.track = track->index;
-            m_status.index = 1; // TODO: handle indexes
+            m_status.index = track->FindIndex(frameAddress);
             m_status.flags = m_status.controlADR == 0x01 ? 0x8 : 0x0;
 
             // TODO: delay seek for a realistic amount of time
@@ -703,7 +705,7 @@ bool CDBlock::SetupGenericPlayback(uint32 startParam, uint32 endParam, uint16 re
                 // Force 1x speed if playing audio track
                 m_targetDriveCycles = kDriveCyclesPlaying1x;
             }
-            devlog::debug<grp::play_init>("Track FAD range {:06X}-{:06X}", m_playStartPos, m_playEndPos);
+            devlog::debug<grp::play_init>("Track:Index FAD range {:06X}-{:06X}", m_playStartPos, m_playEndPos);
 
             if (resetPos) {
                 m_status.frameAddress = m_playStartPos;
@@ -777,7 +779,7 @@ bool CDBlock::SetupFilePlayback(uint32 fileID, uint32 offset, uint8 filterNumber
     m_status.repeatCount = 0; // first repeat
     m_status.controlADR = session.tracks[trackIndex].controlADR;
     m_status.track = trackIndex + 1;
-    m_status.index = 1; // TODO: handle indexes
+    m_status.index = 1;
 
     devlog::debug<grp::play_init>("Read file {}, offset {}, filter {}, frame addresses {:06X} to {:06X}", fileID,
                                   offset, filterNumber, m_playStartPos, m_playEndPos);
@@ -956,7 +958,7 @@ void CDBlock::ProcessDriveStatePlay() {
 
                     m_status.frameAddress++;
                     m_status.track = track->index;
-                    m_status.index = 1; // TODO: handle indexes
+                    m_status.index = track->FindIndex(m_status.frameAddress);
                     m_status.controlADR = track->controlADR;
                     m_status.flags = track->controlADR == 0x41 ? 0x8 : 0x0;
                 }
@@ -1168,7 +1170,7 @@ bool CDBlock::SetupSubcodeTransfer(uint8 type) {
         if (m_disc.sessions.back().tracks[m_status.track - 1].sectorSize < 2448) {
             m_xferBuffer.fill(0xFF);
         } else {
-            devlog::debug<grp::xfer>("Subcode R-W transfer is unimplemented");
+            devlog::trace<grp::xfer>("Subcode R-W transfer is unimplemented");
             m_xferBuffer.fill(0xFF);
         }