+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Core: Performance improvements, especially for ARM builds. (@Wunkolo)
 - Input: Implemented 3D Control Pad. (#28)
 - Media: Preliminary support for CHD files. (#48)
+- Media: Support multi-indexed audio tracks (BIN/CUE only). (#58)
 
 ### Fixes